refactor: move pipeline labels subcommands to separate package
