+++ src/main/java/io/seqera/tower/cli/commands/PipelinesCmd.java
@@ -21,7 +21,7 @@ import io.seqera.tower.cli.commands.pipelines.AddCmd;
 import io.seqera.tower.cli.commands.pipelines.DeleteCmd;
 import io.seqera.tower.cli.commands.pipelines.ExportCmd;
 import io.seqera.tower.cli.commands.pipelines.ImportCmd;
-import io.seqera.tower.cli.commands.pipelines.LabelsCmd;
+import io.seqera.tower.cli.commands.pipelines.labels.LabelsCmd;
 import io.seqera.tower.cli.commands.pipelines.ListCmd;
 import io.seqera.tower.cli.commands.pipelines.UpdateCmd;
 import io.seqera.tower.cli.commands.pipelines.ViewCmd;
