+++ src/main/java/io/seqera/tower/cli/commands/pipelines/labels/PipelinesLabelsManager.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.seqera.tower.cli.commands.pipelines;
+package io.seqera.tower.cli.commands.pipelines.labels;
 
 import java.util.List;
